recent changes, moved models in extra directory
